完善UI；添加游戏细节；添加掉落机制

+++ Assets/Game/Scripts/Enemy/Entity/NoMovementEnemy.cs
@@ -10,7 +10,7 @@ public class NoMovementEnemy : Enemy
         
     }
 
-    public override void TakeDamage(float damage)
+    public override void TakeDamage(int damage)
     {
         base.TakeDamage(damage);
     }
